Found a bug on Downloader::OpenGraph
Turns out when you use the combo:

Down::Http.read() |> ::OpenGraph.new()

It fails randomly when the body of the image is too big,
making the whole process fail.

Changing to a combo of

HTTP.head(src) |> ::OpenGraph.new()

The firs one is just to figure out if the address is responding,
then letting ::OpenGraph work as intended.