Using Transaction object in the ActiveJob
In this last piece we are moving all the logic from PreviewJob
to the already introduced and test Transaction object.

The Transaction class is in charge on how and what we are doing
in order to get the process done, leaving the workflow
more more clear and focused, the other components as
Controller, Model and ActiveJobs has their responsabilities
and are tested according.